Remove `getInProgressSessionDescription` from interface
This does not need to live on the service as it's not directly about chat sessions
